backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
motor==3.6.0
pymongo==4.9.2
httpx==0.28.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.10.2
